fix: rename package and filenames to reflect component name
