+++ component/loki/source/azure_event_hubs/azure_event_hubs_test.go
@@ -1,4 +1,4 @@
-package azureeventhubs
+package azure_event_hubs
 
 import (
 	"testing"
