+++ component/loki/source/azure_event_hubs/internal/parser/parser.go
