+++ component/loki/source/azure_event_hubs/internal/parser/testdata/custom_payload_json_with_records_string.json
