+++ component/loki/source/azure_event_hubs/internal/parser/testdata/custom_payload_text.txt
