+++ component/loki/source/azure_event_hubs/internal/parser/testdata/function_app_logs_message.txt
