+++ component/loki/source/azure_event_hubs/internal/parser/testdata/logic_app_logs_message.json
