+++ component/loki/source/azure_event_hubs/internal/parser/testdata/message_with_invalid_time.json
